Fix bug for TOT calculation over short length
If the strip length is small (<0.5 m) the max time for compute does not
allow the discharge of the preamplifier, therefore the TOT cannot be
computed. This is fixed by assuming a larger max time for compute for
short lengths.
Also add loop to study the behaviout as a function of both charge and
strip length.