Change model input from Jpeak to charge on strip
To have a smoother integration in ATHENA the model should take as input
the charge deposited on strip (C) instead of Jpeak. Changed the input
accordingly. Inside the model I computed Jpeak based on the charge
information, the program itself is not changed. Changed the fitting
scripts in order to use the charge (fC) for the fitting. Changed the TOA
fitting script to show the difference between the fitted curve and the
real data, instead of the surface interpolating the real data. Also
included the intercept in the TOA fit as it will be valid only in the
fitted region.